tests/wpt/tests/css/CSS2/floats/float-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
// Use default transition-behavior: normal.
test_no_interpolation({
  property: 'float',
  from: 'initial',
  to: 'right'
});

test_interpolation({
  property: 'float',
  behavior: 'allow-discrete',
  from: 'left',
  to: 'right',
}, [
  {at: -1, expect: 'left'},
  {at: 0, expect: 'left'},
  {at: 0.4, expect: 'left'},
  {at: 0.5, expect: 'right'},
  {at: 1, expect: 'right'},
  {at: 1.5, expect: 'right'},
]);
</script>
